+++ src/apis/rembrandt/rembrandt.ts
@@ -3,4 +3,5 @@ export * from '@/apis/rembrandt/lib/ResourceType';
 export * from '@/apis/rembrandt/lib/ResourceInstance';
 export * from '@/apis/rembrandt/lib/OptimizationAlgorithm';
 export * from '@/apis/rembrandt/lib/Transformer';
+export * from '@/apis/rembrandt/lib/Recipe';
 export * from '@/apis/rembrandt/lib/transformerTypes';
